+++ CHANGELOG.md
@@ -24,6 +24,22 @@ The following emojis are used to highlight certain changes:
 
 ### Security
 
+## [v0.13.1]
+
+### Added
+
+* An option `DisableHTMLErrors` has been added to `gateway.Config`. When this option
+  is `true`, pretty HTML error pages for web browsers are disabled. Instead, a
+  `text/plain` page with the raw error message as the body is returned.
+
+### Changed
+
+### Removed
+
+### Fixed
+
+### Security
+
 ## [v0.13.0]
 
 ### Added